pyproject.toml
[project]
name = "dungeonmind"
version = "0.2.0"
description = "Foundation API server for DungeonMind projects"
authors = [
    {name = "Alan Meigs", email = "alan.meigs@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.6",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "authlib==1.3.2",
    "boto3==1.35.58",
    "botocore==1.35.99",
    "cachecontrol==0.14.3",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.3",
    "distro==1.9.0",
    "docopt==0.6.2",
    "fal-client==0.5.6",
    "fastapi==0.115.4",
    "filelock==3.18.0",
    "firebase-admin==6.6.0",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "google-api-core==2.24.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.2",
    "google-auth-httplib2==0.2.0",
    "google-cloud-core==2.4.3",
    "google-cloud-firestore==2.19.0",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "hf-xet==1.1.2",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.32.3",
    "idna==3.10",
    "iniconfig==2.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "joblib==1.5.1",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "multidict==6.4.4",
    "networkx==3.5",
    "numpy==2.1.3",
    "openai==1.82.1",
    "packaging==25.0",
    "pandas==2.2.3",
    "pdf2image==1.17.0",
    "pillow==11.2.1",
    "pluggy==1.6.0",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.17",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rsa==4.9.1",
    "s3transfer==0.10.4",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.3",
    "sentence-transformers==3.3.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "torch==2.7.0",
    "tqdm==4.67.1",
    "transformers==4.52.4",
    "triton==3.3.0",
    "twilio==9.6.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.32.0",
    "watchdog==6.0.0",
    "yarl==1.20.0"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-watch==4.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["dungeonmind"]
